Add files via upload
NTPClient library is used to get time from an NTP server which is used to set the time for the Digital Clock. 
NTPClient timeClient(ntpUDP, "pool.ntp.org", utcOffsetInSeconds), the utcOffsetInSeconds must be set according to the users location. For example New York has UTC-05:00 then utcOffsetInSeconds = 60*60*5 = 18000. timeClient.update() along with hours_str = timeClient.getHours(), mins_str = timeClient.getMinutes(), and secs_str = timeClient.getSeconds() is used to get the time variables for the Timer. But the time that we get from them are in String format so we must first convert them into uint8_t data type, which is done by stringToUint() function.